# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: conversations
    description: Conversation operations
  - name: messages
    description: Message operations
  - name: memories
    description: User memory operations
  - name: settings
    description: User settings operations
  - name: profiles
    description: User profile operations
  - name: chat
    description: AI chat operations
  - name: uploads
    description: File upload operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /conversations:
    get:
      operationId: listConversations
      tags: [conversations]
      summary: List user conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
        "401":
          description: Unauthorized
    post:
      operationId: createConversation
      tags: [conversations]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "401":
          description: Unauthorized

  /conversations/{id}:
    patch:
      operationId: updateConversation
      tags: [conversations]
      summary: Update a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteConversation
      tags: [conversations]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /conversations/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
    post:
      operationId: createMessage
      tags: [messages]
      summary: Create a message in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized

  /memories:
    get:
      operationId: listMemories
      tags: [memories]
      summary: List user memories
      responses:
        "200":
          description: List of memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
        "401":
          description: Unauthorized
    post:
      operationId: upsertMemory
      tags: [memories]
      summary: Upsert a memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryInput"
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
        "401":
          description: Unauthorized

  /memories/{key}:
    delete:
      operationId: deleteMemory
      tags: [memories]
      summary: Delete a memory by key
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get user settings
      responses:
        "200":
          description: User settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
        "401":
          description: Unauthorized
    put:
      operationId: upsertSettings
      tags: [settings]
      summary: Upsert user settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSettingsInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
        "401":
          description: Unauthorized

  /profile:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get user profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
    put:
      operationId: upsertProfile
      tags: [profiles]
      summary: Upsert user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized

  /chat:
    post:
      operationId: streamChat
      tags: [chat]
      summary: Stream a chat response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          description: Unauthorized

  /chat/title:
    post:
      operationId: generateTitle
      tags: [chat]
      summary: Generate a conversation title
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TitleRequest"
      responses:
        "200":
          description: Title
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TitleResponse"
        "401":
          description: Unauthorized

  /chat/image:
    post:
      operationId: generateImage
      tags: [chat]
      summary: Generate an image from a prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageRequest"
      responses:
        "200":
          description: Image URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageResponse"
        "401":
          description: Unauthorized

  /chat/logo:
    post:
      operationId: generateLogo
      tags: [chat]
      summary: Generate an SVG logo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoRequest"
      responses:
        "200":
          description: SVG logo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoResponse"
        "401":
          description: Unauthorized


components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Conversation:
      type: object
      required: [id, userId, title, modelId, pinned, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        modelId:
          type: string
        pinned:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string

    ConversationInput:
      type: object
      required: [title, modelId]
      properties:
        title:
          type: string
        modelId:
          type: string

    ConversationUpdate:
      type: object
      properties:
        title:
          type: string
        modelId:
          type: string
        pinned:
          type: boolean
        updatedAt:
          type: string

    Message:
      type: object
      required: [id, conversationId, userId, role, content, createdAt]
      properties:
        id:
          type: string
        conversationId:
          type: string
        userId:
          type: string
        role:
          type: string
        content:
          type: string
        modelId:
          type: ["string", "null"]
        attachments:
          type: ["array", "null"]
          items:
            type: string
        createdAt:
          type: string

    MessageInput:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
        content:
          type: string
        modelId:
          type: string
        attachments:
          type: array
          items:
            type: string

    Memory:
      type: object
      required: [id, userId, key, value, weight]
      properties:
        id:
          type: string
        userId:
          type: string
        key:
          type: string
        value:
          type: string
        weight:
          type: integer

    MemoryInput:
      type: object
      required: [key, value]
      properties:
        key:
          type: string
        value:
          type: string
        weight:
          type: integer

    UserSettings:
      type: object
      required: [userId, persona, language, responseStyle, responseLength, animationSpeed, demoMode, adultMode, soundEnabled, voiceEnabled]
      properties:
        userId:
          type: string
        persona:
          type: string
        language:
          type: string
        responseStyle:
          type: string
        responseLength:
          type: string
        animationSpeed:
          type: string
        demoMode:
          type: boolean
        adultMode:
          type: boolean
        soundEnabled:
          type: boolean
        voiceEnabled:
          type: boolean
        permissions:
          type: object

    UserSettingsInput:
      type: object
      properties:
        persona:
          type: string
        language:
          type: string
        responseStyle:
          type: string
        responseLength:
          type: string
        animationSpeed:
          type: string
        demoMode:
          type: boolean
        adultMode:
          type: boolean
        soundEnabled:
          type: boolean
        voiceEnabled:
          type: boolean
        permissions:
          type: object

    Profile:
      type: object
      required: [id, displayName, mascotPersonality]
      properties:
        id:
          type: string
        displayName:
          type: string
        mascotPersonality:
          type: string
        email:
          type: ["string", "null"]

    ProfileInput:
      type: object
      properties:
        displayName:
          type: string
        mascotPersonality:
          type: string
        email:
          type: string

    ChatRequest:
      type: object
      required: [messages, modelId]
      properties:
        messages:
          type: array
          items:
            type: object
            required: [role, content]
            properties:
              role:
                type: string
              content:
                type: string
        modelId:
          type: string
        attachments:
          type: array
          items:
            type: string
        memories:
          type: array
          items:
            type: object
        settings:
          type: object

    TitleRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string

    TitleResponse:
      type: object
      required: [title]
      properties:
        title:
          type: string

    ImageRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string

    ImageResponse:
      type: object
      required: [image]
      properties:
        image:
          type: string
        finalPrompt:
          type: string

    LogoRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string

    LogoResponse:
      type: object
      required: [svg]
      properties:
        svg:
          type: string