# .migration-backup/lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: string;
  conversationId: string;
  userId: string;
  role: string;
  content: string;
  /** @nullable */
  modelId?: string | null;
  /** @nullable */
  attachments?: string[] | null;
  createdAt: string;
}