# .migration-backup/lib/api-zod/src/generated/types/userSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserSettingsPermissions } from './userSettingsPermissions';

export interface UserSettings {
  userId: string;
  persona: string;
  language: string;
  responseStyle: string;
  responseLength: string;
  animationSpeed: string;
  demoMode: boolean;
  adultMode: boolean;
  soundEnabled: boolean;
  voiceEnabled: boolean;
  permissions?: UserSettingsPermissions;
}